move logic into a custom hook
